feat(mcp): extract OAuth and bearer auth into lilith-mcp 0.2.0
- packages/mcp 0.2.0: parameterized OAuth 2.1 endpoints (AS metadata,
  RFC 9728 resource metadata, dynamic client registration, authorize,
  token) and an access-token verifier driven by a per-package
  OAuthConfig; createMcpAuth bundles the bearer-first request context,
  scope tracking, requireScope, and 401 discovery headers in one
  closure so scope bookkeeping stays consistent.
- packages/readr: oauth.ts reduced to its config block plus
  re-exports; mcp.ts delegates all auth wiring to the shared module;
  keystone.ts unchanged. Behavior verified identical by the full local
  regression matrix (cookie, OAuth PKCE flow, scope denial, discovery
  headers).

Adopting packages now supply only a config block; protocol and auth
logic stay in one audited place.
