+++ packages/mcp/src/index.ts
@@ -10,3 +10,19 @@ export type {
   McpTool,
   RequestContextFactory,
 } from './types'
+export {
+  createAccessTokenVerifier,
+  createOAuthEndpoints,
+  isOAuthConfigured,
+} from './oauth'
+export type {
+  AccessTokenClaims,
+  OAuthCommonContext,
+  OAuthConfig,
+  OAuthKeystoneContext,
+  OAuthNext,
+  OAuthRequest,
+  OAuthResponse,
+} from './oauth'
+export { createMcpAuth } from './mcp-auth'
+export type { BearerAuthContext } from './mcp-auth'
