+++ packages/readr/package.json
@@ -21,7 +21,7 @@
     "@keystone-6/auth": "7.0.0",
     "@keystone-6/core": "5.2.0",
     "@mirrormedia/lilith-core": "^3.0.1",
-    "@mirrormedia/lilith-mcp": "0.1.0",
+    "@mirrormedia/lilith-mcp": "0.2.0",
     "express": "^4.17.1",
     "graphql-upload": "^15.0.2",
     "http-proxy-middleware": "^2.0.3",
